requirements.txt
google-cloud-vision
flask~=2.0.1
